switched to 32bit R/C

+++ Core/Libs/Common/Common.c
@@ -12,10 +12,10 @@ uint16_t Common_Power(uint16_t base, uint16_t exp) {
 	}
 }
 
-void Common_GetSignificantAndMantissa(uint16_t number, uint16_t base, uint8_t digits, uint16_t *significant, uint16_t *mantissa) {
+void Common_GetSignificantAndMantissa(uint32_t number, uint16_t base, uint8_t digits, uint16_t *significant, uint16_t *mantissa) {
 	*significant = number / base;
 
-	uint16_t remainder = number % base;
+	uint32_t remainder = number % base;
 
 	// add half base for rounding
 	*mantissa = (remainder * Common_Power(10, digits) + base / 2) / base;
